fix sdr_transceiver_ft8/app/stop.sh

+++ projects/sdr_transceiver_ft8/app/stop.sh
@@ -1,7 +1,8 @@
 #! /bin/sh
 
-rm -f /etc/periodic/ft8
-
-service dcron restart
-
-killall -q decode-ft8.sh write-c2-files ft8d
+if test -e /etc/periodic/ft8
+then
+  rm -f /etc/periodic/ft8
+  service dcron restart
+  killall -q decode-ft8.sh write-c2-files ft8d
+fi
